+++ static/app/assets/AmbientManager-BAEkHO0Q.js
@@ -1,0 +1,1 @@
+import{r as e}from"./index-CW4j7rui.js";export{e as ambient};